Move the Debian-related notes to dodge dpkg err
With the packaging notes placed into the main README file,
`dpkg-buildpacakge` complains that the upstream release differs from the
currently-active project.

Which is true: The upstream version I said I'd build is tag
`upstream/0.0_git20250101.d0bd3ff6`, but the version I'm actually
building contains those extra notes in the README.

I'm moving the notes into the `debian/` folder so that it doesn't
collide with this problem.